Refactor tests for API Service
