+++ src/webclient/tests/api-service.js
@@ -1,10 +1,10 @@
-import ApiService from '../main/services/api-secrvice';
+import APIService from '../main/services/api-service';
 import UserData from '../models/user-data';
 import AuthorizationError from '../models/errors/authorization-error';
 import fetchMock from '../../../node_modules/fetch-mock/esm/client.js';
 
 const {module, test} = QUnit;
-const service = new ApiService(false);
+const service = new APIService(false);
 fetchMock.config.overwriteRoutes = true;
 
 export default module('App service test', function(hook) {
@@ -30,9 +30,7 @@ export default module('App service test', function(hook) {
       return 200;
     })));
     const userData = new UserData('Alex1', 'Mdaskjdsdasa1543');
-    service.register(userData).catch((error) => {
-      assert.ok(true, 'Should return exception if user with this login already registered.');
-    });
+    assert.rejects(service.register(userData), 'Should return exception if user with this login already registered.');
   });
 
   test('Login method should return exception if user with this login is not registered.', async (assert) => {
@@ -45,8 +43,6 @@ export default module('App service test', function(hook) {
       return 200;
     })));
     const userData = new UserData('Vas9', 'Mdaskjdsdasa1543');
-    await service.logIn(userData).catch((error) => {
-      assert.ok(error instanceof AuthorizationError, 'Should throw AuthorizationError.');
-    });
+    assert.rejects(service.logIn(userData), 'Should return exception if user with this login is not registered.');
   });
 });
